added tests on incorrect login/password

See merge request qa-produit/tests/qa_tina_tests!720

+++ qa_tina_tests/USER/API/ICU/Access_Keys/test_DeleteAccessKey.py
@@ -2,7 +2,7 @@ from time import sleep
 
 from qa_sdk_common.exceptions.osc_exceptions import OscApiException
 from qa_sdk_pub import osc_api
-from qa_test_tools.misc import id_generator, assert_error
+from qa_test_tools import misc
 from qa_tina_tools.test_base import OscTinaTest
 
 
@@ -23,7 +23,7 @@ class Test_DeleteAccessKey(OscTinaTest):
             self.a1_r1.icu.DeleteAccessKey()
             assert False, "Call should not have been successful"
         except OscApiException as error:
-            assert_error(error, 400, 'IcuClientException', 'Field AccessKeyId is required')
+            misc.assert_error(error, 400, 'IcuClientException', 'Field AccessKeyId is required')
 
     def test_T346_with_param_but_empty(self):
         sleep(30)
@@ -31,7 +31,7 @@ class Test_DeleteAccessKey(OscTinaTest):
             self.a1_r1.icu.DeleteAccessKey(AccessKeyId='')
             assert False, "Call should not have been successful"
         except OscApiException as error:
-            assert_error(error, 400, 'MissingParameter', 'Parameter cannot be empty: AccessKeyId')
+            misc.assert_error(error, 400, 'MissingParameter', 'Parameter cannot be empty: AccessKeyId')
 
     def test_T1799_empty_param_with_more_than_one_existing(self):
         sleep(30)
@@ -42,7 +42,7 @@ class Test_DeleteAccessKey(OscTinaTest):
             self.a1_r1.icu.DeleteAccessKey(AccessKeyId='')
             assert False, "Call should not have been successful"
         except OscApiException as error:
-            assert_error(error, 400, 'MissingParameter', 'Parameter cannot be empty: AccessKeyId')
+            misc.assert_error(error, 400, 'MissingParameter', 'Parameter cannot be empty: AccessKeyId')
         finally:
             if ret_create:
                 sleep(30)
@@ -54,19 +54,20 @@ class Test_DeleteAccessKey(OscTinaTest):
             self.a1_r1.icu.DeleteAccessKey(AccessKeyId='foo')
             assert False, "Exception should have been raised"
         except OscApiException as error:
-            assert_error(error, 400, 'NoSuchEntity', None)
+            misc.assert_error(error, 400, 'NoSuchEntity', None)
 
     def test_T348_invalid_accesskeyid_validAK_not_existant(self):
         sleep(30)
         try:
-            ak = id_generator(size=20)
+            ak = misc.id_generator(size=20)
             self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
             assert False, "Exception should have been raised"
         except OscApiException as error:
-            assert_error(error, 400, 'NoSuchEntity', None)
+            misc.assert_error(error, 400, 'NoSuchEntity', None)
 
     def test_T349_invalid_accesskeyid_AK_partially_existant(self):
         sleep(30)
+        ak = None
         ret_create = None
         ret_delete = None
         try:
@@ -76,14 +77,15 @@ class Test_DeleteAccessKey(OscTinaTest):
             ret_delete = self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak_modified)
             assert False, "Call should not have been successful"
         except OscApiException as error:
-            assert_error(error, 400, 'NoSuchEntity', None)
+            misc.assert_error(error, 400, 'NoSuchEntity', None)
         finally:
-            if ret_create and not ret_delete:
+            if ret_create and ak and not ret_delete:
                 sleep(30)
                 self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
 
     def test_T350_valid_accesskeyid(self):
         sleep(30)
+        ak = None
         ret_create = None
         ret_delete = None
         try:
@@ -92,12 +94,13 @@ class Test_DeleteAccessKey(OscTinaTest):
             ret_delete = self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
             assert ret_delete.response
         finally:
-            if ret_create and not ret_delete:
+            if ret_create and ak and not ret_delete:
                 sleep(30)
                 self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
 
     def test_T3973_method_authAkSk(self):
         sleep(30)
+        ak = None
         ret_create = None
         ret_delete = None
         try:
@@ -106,12 +109,13 @@ class Test_DeleteAccessKey(OscTinaTest):
             ret_delete = self.a1_r1.icu.DeleteAccessKey(exec_data={osc_api.EXEC_DATA_AUTHENTICATION: osc_api.AuthMethod.AkSk}, AccessKeyId=ak)
             assert ret_delete.response
         finally:
-            if ret_create and not ret_delete:
+            if ret_create and ak and not ret_delete:
                 sleep(30)
                 self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
 
     def test_T3974_with_method_authLoginPassword(self):
         sleep(30)
+        ak = None
         ret_create = None
         ret_delete = None
         try:
@@ -121,12 +125,32 @@ class Test_DeleteAccessKey(OscTinaTest):
                                                         AccessKeyId=ak)
             assert ret_delete.response
         finally:
-            if ret_create and not ret_delete:
+            if ret_create and ak and not ret_delete:
+                sleep(30)
+                self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
+
+    def test_T5988_with_method_authLoginPassword_incorrect(self):
+        sleep(30)
+        ak = None
+        ret_create = None
+        ret_delete = None
+        try:
+            ret_create = self.a1_r1.icu.CreateAccessKey()
+            ak = ret_create.response.accessKey.accessKeyId
+            self.a1_r1.icu.DeleteAccessKey(exec_data={osc_api.EXEC_DATA_AUTHENTICATION: osc_api.AuthMethod.LoginPassword,
+                                                      osc_api.EXEC_DATA_LOGIN: 'foo', osc_api.EXEC_DATA_PASSWORD: 'bar'},
+                                                      AccessKeyId=ak)
+            assert False, 'Call should not have been successful'
+        except OscApiException as error:
+            misc.assert_error(error, 403, 'InvalidLoginPassword', 'Account foo failed to authenticate.')
+        finally:
+            if ret_create and ak and not ret_delete:
                 sleep(30)
                 self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
 
     def test_T351_valid_accesskeyid_other_account(self):
         sleep(30)
+        ak = None
         ret_create = None
         ret_delete = None
         try:
@@ -135,9 +159,9 @@ class Test_DeleteAccessKey(OscTinaTest):
             ret_delete = self.a2_r1.icu.DeleteAccessKey(AccessKeyId=ak)
             assert False, "Call should not have been successful"
         except OscApiException as error:
-            assert_error(error, 400, 'NoSuchEntity', None)
+            misc.assert_error(error, 400, 'NoSuchEntity', None)
         finally:
-            if ret_create and not ret_delete:
+            if ret_create and ak and not ret_delete:
                 self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
 
     def test_T3773_check_throttling(self):
@@ -165,6 +189,7 @@ class Test_DeleteAccessKey(OscTinaTest):
 
     def test_T5744_with_extra_param(self):
         sleep(30)
+        ak = None
         ret_create = None
         ret_delete = None
         try:
@@ -173,6 +198,6 @@ class Test_DeleteAccessKey(OscTinaTest):
             ret_delete = self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak, Foo='Bar')
             assert ret_delete.response
         finally:
-            if ret_create and not ret_delete:
+            if ret_create and ak and not ret_delete:
                 sleep(30)
                 self.a1_r1.icu.DeleteAccessKey(AccessKeyId=ak)
